[MOB-32] Renaming domain errors
