+++ src/domain/errors/api-error.ts
